+++ src/app/services/mock-backend.js
@@ -986,12 +986,12 @@ export default Service.extend({
     });
     const archiveRecallInfo = store.createRecord('archive-recall-info', {
       id: infoGri,
-      sourceArchive: archive,
-      sourceDataset: await get(archive, 'dataset'),
-      targetFiles: 100,
-      targetBytes: 1000000,
-      startTimestamp: null,
-      finishTimestamp: null,
+      archive,
+      dataset: await get(archive, 'dataset'),
+      totalFileCount: 100,
+      totalByteSize: 1000000,
+      startTime: null,
+      finishTime: null,
     });
     const archiveRecallState = store.createRecord('archive-recall-state', {
       id: stateGri,
@@ -1031,57 +1031,57 @@ export default Service.extend({
     const archiveRecallInfo = this.get('entityRecords.archiveRecallInfo.0');
     const archiveRecallState = this.get('entityRecords.archiveRecallState.0');
     const {
-      startTimestamp,
-      finishTimestamp,
-      targetFiles,
-      targetBytes,
+      startTime,
+      finishTime,
+      totalFileCount,
+      totalByteSize,
     } = getProperties(
       archiveRecallInfo,
-      'startTimestamp',
-      'finishTimestamp',
-      'targetFiles',
-      'targetBytes',
+      'startTime',
+      'finishTime',
+      'totalFileCount',
+      'totalByteSize',
     );
-    if (finishTimestamp) {
+    if (finishTime) {
       return;
     }
 
     let infoModified = false;
     let {
-      currentBytes,
-      currentFiles,
+      bytesCopied,
+      filesCopied,
     } = getProperties(
       archiveRecallState,
-      'currentBytes',
-      'currentFiles',
+      'bytesCopied',
+      'filesCopied',
     );
-    if (!startTimestamp) {
-      set(archiveRecallInfo, 'startTimestamp', Date.now());
+    if (!startTime) {
+      set(archiveRecallInfo, 'startTime', Date.now());
       infoModified = true;
     }
-    if (currentBytes < targetBytes) {
-      const filesIncrement = Math.floor(targetFiles / stepsCount);
-      const bytesIncrement = Math.floor(targetBytes / stepsCount);
-      currentFiles = Math.min(targetFiles, currentFiles + filesIncrement);
-      currentBytes = Math.min(targetBytes, currentBytes + bytesIncrement);
+    if (bytesCopied < totalByteSize) {
+      const filesIncrement = Math.floor(totalFileCount / stepsCount);
+      const bytesIncrement = Math.floor(totalByteSize / stepsCount);
+      filesCopied = Math.min(totalFileCount, filesCopied + filesIncrement);
+      bytesCopied = Math.min(totalByteSize, bytesCopied + bytesIncrement);
     }
     setProperties(archiveRecallState, {
-      currentBytes,
-      currentFiles,
+      bytesCopied,
+      filesCopied,
     });
 
     const savePromises = [];
-    const finished = currentBytes >= targetBytes || currentFiles >= targetFiles;
+    const finished = bytesCopied >= totalByteSize || filesCopied >= totalFileCount;
     if (finished) {
       // just to be certain
-      currentFiles = targetFiles;
-      currentBytes = targetBytes;
-      set(archiveRecallInfo, 'finishTimestamp', Date.now());
+      filesCopied = totalFileCount;
+      bytesCopied = totalByteSize;
+      set(archiveRecallInfo, 'finishTime', Date.now());
       infoModified = true;
     }
     setProperties(archiveRecallState, {
-      currentBytes,
-      currentFiles,
+      bytesCopied,
+      filesCopied,
     });
     if (infoModified) {
       savePromises.push(archiveRecallInfo.save());